refactor: Product 도메인 enums 패키지 평면화
