+++ src/main/java/kr/hhplus/be/server/domain/product/Product.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.domain.product;
 
 import jakarta.persistence.*;
-import kr.hhplus.be.server.domain.product.enums.Category;
 import kr.hhplus.be.server.support.entity.BaseEntity;
 import kr.hhplus.be.server.support.exception.ApiErrorCode;
 import kr.hhplus.be.server.support.exception.ApiException;
