+++ src/main/java/kr/hhplus/be/server/domain/product/ProductSalesRankInfo.java
@@ -1,7 +1,5 @@
 package kr.hhplus.be.server.domain.product;
 
-import kr.hhplus.be.server.domain.product.enums.RankingScope;
-
 public record ProductSalesRankInfo(
         Long productId,
         int totalSalesCount,
