+++ src/main/java/kr/hhplus/be/server/domain/product/ProductSalesRankRepository.java
@@ -1,7 +1,5 @@
 package kr.hhplus.be.server.domain.product;
 
-import kr.hhplus.be.server.domain.product.enums.RankingScope;
-
 import java.util.List;
 
 public interface ProductSalesRankRepository {
