+++ src/main/java/kr/hhplus/be/server/infra/product/ProductSalesRankJpaRepository.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.infra.product;
 
 import kr.hhplus.be.server.domain.product.ProductSalesRank;
-import kr.hhplus.be.server.domain.product.enums.RankingScope;
+import kr.hhplus.be.server.domain.product.RankingScope;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
