+++ src/main/java/kr/hhplus/be/server/infra/product/ProductSalesRankRepositoryImpl.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.infra.product;
 
 import kr.hhplus.be.server.domain.product.ProductSalesRank;
 import kr.hhplus.be.server.domain.product.ProductSalesRankRepository;
-import kr.hhplus.be.server.domain.product.enums.RankingScope;
+import kr.hhplus.be.server.domain.product.RankingScope;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
